% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_applicants.R
\name{assign_cohort}
\alias{assign_cohort}
\title{Assign a batch of applicants}
\usage{
assign_cohort(applicants)
}
\description{
Take a batch of raw applicant data, process it for assignment
Draw assignments, and upload them to the google drive.
}
