% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_exact.R
\name{assign_to_condition}
\alias{assign_to_condition}
\title{Assign a table of applicants imported with read_applicant_file()}
\usage{
assign_to_condition(
  applicants,
  n_offers_by_program = list(),
  seed,
  browse = FALSE
)
}
\arguments{
\item{n_offers_by_program}{Controls the program-specific number of offers for the cohort of applicants
being assigned, in the form \verb{list(<program_short> = <number_of_offers>)}.
For example, in the first assignment round, we were required to assign 217
JITA applicants and 155 JDA applicants to the treatment group. In this case,
\code{n_offers_by_program = list(JITA = 217, JDA = 155)}}
}
\description{
Assign a table of applicants imported with read_applicant_file()
}
