% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default experiment parameters}
\format{
a list with some number of named values
}
\usage{
default_parameters
}
\description{
A list of default parameters to be used in stratification and randomization
}
\keyword{datasets}
