% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignment_list.R
\name{generate_assignment_block}
\alias{generate_assignment_block}
\title{Generate block for assignment list}
\usage{
generate_assignment_block(target_treatment_probability, block_size)
}
\description{
Make a block of assignments of a given size
with a given treatment probability
}
