% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignment_list.R
\name{generate_assignment_list}
\alias{generate_assignment_list}
\title{Generate blank binary assignment list}
\usage{
generate_assignment_list(
  minimum_list_size,
  target_treatment_probability,
  block_size
)
}
\description{
This uses block randomization to generate a blank list of assignments
of at least a given size with a treatment probability as close as possible
to the target
}
