% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignment_list.R
\name{generate_assignment_list_bundle}
\alias{generate_assignment_list_bundle}
\title{Generate a bundle of assignment lists}
\usage{
generate_assignment_list_bundle(parameters = list())
}
\description{
Create assignment lists with a range of treatment
probabilities, whose incrememts are determined by the
block size.
}
