% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_stratification_parameters}
\alias{get_latest_stratification_parameters}
\title{Retrieve latest version of stratification parameters}
\usage{
get_latest_stratification_parameters()
}
\description{
Retrieve latest version of stratification parameters
}
