% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_experiment.R
\name{initialize_experiment}
\alias{initialize_experiment}
\title{Initialize the NPower RCT}
\usage{
initialize_experiment(custom_parameters = list(), test = FALSE)
}
\arguments{
\item{custom_parameters}{A list of parameters to update, from among the following:
\itemize{
\item minimum_list_size = 5000,
\item minimum_treatment_probability = 0.2,
\item maximum_treatment_probability = 0.7,
\item block_size = 20,
\item rep_target_gender_diversity = 0.5,
\item eligible_programs = "JITA",
\item eligible_locations = "Toronto;Peel;York"
}}
}
\description{
Creates three googlesheets in the npowerassginmentdatalink
google drive:
\itemize{
\item activity-log: tracks activities, including assignment list creation,
assignment batch execution, and stratification parameter updates.
\item stratification-parameters: records initial experiment parameters and
parameter updates.
\item assignment-list/assignment-list_\link{timestamp}
}
}
