% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_exact.R
\name{read_applicant_file}
\alias{read_applicant_file}
\title{Import and clean up a report file of applicants from NPower.}
\usage{
read_applicant_file(path, quiet = FALSE)
}
\description{
Import and clean up a report file of applicants from NPower.
}
