% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_cohort}
\alias{synthetic_cohort}
\title{Sythetic cohort of 500 NPower applicants from various programs}
\format{
A tibble with 500 rows and 5 variables:
\describe{
\item{applicant_id}{unique identifier for each applicant}
\item{gender}{applicant gender, from \link{'Man', 'Woman', 'Other'}}
\item{assignment_eligible}{1 if the applicant should be assigned}
\item{location}{delivery location}
\item{program}{program applied to}
}
}
\usage{
synthetic_cohort
}
\description{
A dataset with synthetic, randomly generated applicants whose
characteristics are chosen to roughly represent the applicants
we will be randomly assigning
}
\keyword{datasets}
