% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_stratification_parameters}
\alias{update_stratification_parameters}
\title{Modify active stratification parameters}
\usage{
update_stratification_parameters(parameter_updates = list(), test = FALSE)
}
\description{
Replace specific stratification parameters with
}
