% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_assignment_list.R
\name{upload_assignment_list_bundle}
\alias{upload_assignment_list_bundle}
\title{Upload assignment list bundle to Google Drive}
\usage{
upload_assignment_list_bundle(
  assignment_list_bundle,
  custom_tag = NULL,
  test = FALSE
)
}
\description{
Function to create a new assignment list workbook
in google sheets and log its creation in the activity log
}
